requirements.txt
Django==5.0.2
Pillow==11.0.0
django-infinite-scroll-pagination==1.1.0
python-dotenv==1.0.1
django-allauth==0.61.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
python-dateutil==2.8.2 